// src/lib/database.types.ts
export type Database = {
  public: {
    Tables: {
// ...
      flashcards: {
        Row: {
          back_content: string
          back_content_html: string | null
          chapter_id: string | null
          character_id: string | null
          created_at: string | null
          difficulty_level: number | null
          front_content: string
          front_content_html: string | null
          hint: string | null
          id: string
          section_id: string | null
          serial_number: number
          updated_at: string | null
        }
        Insert: {
          back_content: string
          back_content_html?: string | null
          chapter_id?: string | null
          character_id?: string | null
          created_at?: string | null
          difficulty_level?: number | null
          front_content: string
          front_content_html?: string | null
          hint?: string | null
          id?: string
          section_id?: string | null
          serial_number: number
          updated_at?: string | null
        }
        Update: {
          back_content?: string
          back_content_html?: string | null
          chapter_id?: string | null
          character_id?: string | null
          created_at?: string | null
          difficulty_level?: number | null
          front_content?: string
          front_content_html?: string | null
          hint?: string | null
          id?: string
          section_id?: string | null
          serial_number?: number
          updated_at?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "flashcards_chapter_id_fkey"
            columns: ["chapter_id"]
            isOneToOne: false
            referencedRelation: "chapters"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "flashcards_character_id_fkey"
            columns: ["character_id"]
            isOneToOne: false
            referencedRelation: "character_details"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "flashcards_character_id_fkey"
            columns: ["character_id"]
            isOneToOne: false
            referencedRelation: "characters"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "flashcards_section_id_fkey"
            columns: ["section_id"]
            isOneToOne: false
            referencedRelation: "sections"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
    }
// ...
  }
}